% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OriginalAux.R, R/RcppExports.R
\name{RaoBlackwellAuxSMC_witcovs_andNA}
\alias{RaoBlackwellAuxSMC_witcovs_andNA}
\alias{conditioned_C_RaoBlackwellAuxSMC_witcovs_andNA}
\title{Sample nSim MC samples from a SUN with time constant terms using Rcpp and Armadillo.}
\usage{
RaoBlackwellAuxSMC_witcovs_andNA(
  y,
  delta,
  gamma,
  x,
  Z,
  matFF,
  mu0,
  Sigma0 = Sigma0,
  SigmaEps = SigmaEps,
  GG = Gt,
  seed = 123,
  nSim = nSim
)

conditioned_C_RaoBlackwellAuxSMC_witcovs_andNA(
  y,
  delta,
  gamma,
  x,
  Z,
  matFF,
  G,
  mu0,
  Sigma0,
  SigmaEps,
  nSim = 100L,
  seed = 1L
)
}
\arguments{
\item{y}{Binary vector of 1,0 or NA}

\item{delta}{Vector for time constant covariates.}

\item{gamma}{Vector for time varying covariates.}

\item{x}{Vector encoding time constant covariates.}

\item{Z}{Matrix encoding time varing covariates, columns are observed times.}

\item{matFF}{Encoding of evolutions coefficents.}

\item{mu0}{Prior mean for starting time.}

\item{Sigma0}{Prior variance for starting time.}

\item{SigmaEps}{Evolution prior.}

\item{seed}{seed to be fixed.}

\item{nSim}{Number of samples.}

\item{G}{Evolution matrix.}
}
\value{
Random samples of evoution coefficents.
}
\description{
Function Sample nSim MC samples from a SUN with time constant terms using Rcpp and Armadillo.
}
\section{Fields}{

\describe{
\item{\code{conditioned_C_RaoBlackwellAuxSMC_witcovs_andNA}}{Rcpp Function.}

\item{\code{RaoBlackwellAuxSMC_witcovs_andNA}}{Original R function.}
}}

\examples{
1+1 == 2
}
