% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R
\docType{package}
\name{HEP-package}
\alias{HEP-package}
\alias{HEP}
\title{Bayesian Non-Parametric clustering of Non-Aligned Time Series}
\description{
The HEP package allows for clustering binary time series data in R.
It provides a comprehensive set of functions and algorithms tailored specifically
for the uniquechallenges posed by binary time series datasets.
}
\details{
A Fantastic R Package

This R package provides incredible functionality to accomplish extraordinary tasks.
}
\examples{
# Esempi di utilizzo del pacchetto
dati <- data.frame(x = 1:10, y = rnorm(10))
2*2

}
\author{
Beatrice Cantoni e Giovanni Poli
}
