% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvrnormArma}
\alias{mvrnormArma}
\alias{mvrnormArma1}
\title{Samples from mutivariate normal using Rcpp and Armadillo}
\usage{
mvrnormArma1(mu, sigma)
}
\arguments{
\item{mu}{Mean vector.}

\item{sigma}{Covariance matrix vector.}

\item{n}{number of samples .}
}
\value{
Random samples
}
\description{
Function to return 1 or n samples from a mutivariate normal using Rcpp and Armadillo
}
\section{Fields}{

\describe{
\item{\code{mvrnormArma}}{Return n sample from a mutivariate normal.}

\item{\code{mvrnormArma1}}{Return a sample from a mutivariate normal.}
}}

\examples{
mvrnormArma(n = 1, mu = c(0,0), sigma = diag(c(1,1)))
}
